Change and Add Columns in ingredients
In order to lay the groundwork for calculating ingredient quantities for lists of recipes, we need to store the ingredient quantity as a numerical value. Here the column _quantity_ is changed from type string to type decimal. Decimal was chosen as this allows for an accurate accounting of decimal values.

We also store a string representation of the quantity as it is presented in the recipes seed data. Here the column name is changed to _quantity_str_. In the future, it may be desirable to programmatically produce these strings from the decimal value as opposed to storing both.